dev-requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file dev-requirements.txt dev-requirements.in
#
argh==0.26.2              # via watchdog
blinker==1.4              # via raven
braintree==3.29.2
click==6.6                # via flask, pip-tools
contextlib2==0.5.4        # via raven
dj-static==0.0.6
django-registration-redux==1.4
git+git://github.com/simon-weber/django-supervisor.git@autoreload_on_created#egg=django-supervisor==0.3.3.dev0
Django==1.8.14
djmail==0.13.0
dnspython==1.14.0
docopt==0.6.2
enum34==1.1.6
first==2.0.1              # via pip-tools
Flask==0.11.1             # via raven
funcsigs==1.0.2           # via mock
futures==3.0.5
gevent==1.1.2
google-api-python-client==1.5.3
google-measurement-protocol==0.1.6
greenlet==0.4.10          # via gevent
gunicorn==19.6.0
httplib2==0.9.2           # via google-api-python-client, oauth2client
itsdangerous==0.24        # via flask
Jinja2==2.8               # via flask
MarkupSafe==0.23          # via jinja2
meld3==1.0.2              # via supervisor
mock==1.3.0
newrelic==2.70.0.51
oauth2client==3.0.0
pathtools==0.1.2          # via watchdog
pbr==1.10.0               # via mock
git+https://github.com/venmo/pip-tools.git@1.7.1a1#egg=pip-tools==1.7.1a1
pyasn1-modules==0.0.8     # via oauth2client
pyasn1==0.1.9             # via oauth2client, pyasn1-modules, rsa
PyYAML==3.12              # via watchdog
raven[flask]==5.26.0
requests==2.11.1
rsa==3.4.2                # via oauth2client
simplejson==3.8.2         # via uritemplate
six==1.10.0               # via google-api-python-client, mock, oauth2client, pip-tools
git+https://github.com/simon-weber/SleekXMPP.git@1.5.0rc1#egg=sleekxmpp==1.5.0rc1
sparkpost==1.2.0
static3==0.7.0            # via dj-static
supervisor==3.3.1
uritemplate==0.6          # via google-api-python-client
watchdog==0.8.3           # via django-supervisor
Werkzeug==0.11.11         # via flask
